reading json is fixed
