+++ resolution/ABI/CNS/cnsRegistry.json
