--- resolution/NamingServices/CNS_ABI/RegistryType.swift
@@ -1,9 +1,0 @@
-// RegistryType.swift
-
-import Foundation
-
-enum RegistryType: String, Codable {
-    case constructor = "constructor"
-    case event = "event"
-    case function = "function"
-}
