--- resolution/NamingServices/CNS_ABI/ResolverType.swift
@@ -1,9 +1,0 @@
-// ResolverType.swift
-
-import Foundation
-
-enum ResolverType: String, Codable {
-    case constructor = "constructor"
-    case event = "event"
-    case function = "function"
-}
