--- resolution/Put.swift
@@ -1,57 +1,0 @@
-// Put.swift
-
-// This file was generated from JSON Schema using quicktype, do not modify it directly.
-// To parse the JSON, add this file to your project and do:
-//
-//   let put = try Put(json)
-
-import Foundation
-
-// MARK: - Put
-struct Put: Codable {
-    let internalType: InternalTypeEnum
-    let name: String
-    let type: InternalTypeEnum
-    let indexed: String?
-}
-
-// MARK: Put convenience initializers and mutators
-
-extension Put {
-    init(data: Data) throws {
-        self = try newJSONDecoder().decode(Put.self, from: data)
-    }
-
-    init(_ json: String, using encoding: String.Encoding = .utf8) throws {
-        guard let data = json.data(using: encoding) else {
-            throw NSError(domain: "JSONDecoding", code: 0, userInfo: nil)
-        }
-        try self.init(data: data)
-    }
-
-    init(fromURL url: URL) throws {
-        try self.init(data: try Data(contentsOf: url))
-    }
-
-    func with(
-        internalType: InternalTypeEnum? = nil,
-        name: String? = nil,
-        type: InternalTypeEnum? = nil,
-        indexed: String?? = nil
-    ) -> Put {
-        return Put(
-            internalType: internalType ?? self.internalType,
-            name: name ?? self.name,
-            type: type ?? self.type,
-            indexed: indexed ?? self.indexed
-        )
-    }
-
-    func jsonData() throws -> Data {
-        return try newJSONEncoder().encode(self)
-    }
-
-    func jsonString(encoding: String.Encoding = .utf8) throws -> String? {
-        return String(data: try self.jsonData(), encoding: encoding)
-    }
-}
